Update vector_rotate.py
Add import statement for Numpy, in order to prepare for using functions of Numpy.
Add argument which is same as vector_rotate.pro, in order to prepare computation using arguments.
Add processing for preparing sin and cos values for the rotation angle theta, in order to prepare computation about rotation matrix.