Update pwe_efd.py
Add options for '*dsi' Tplot Variables of datatype = 'Espin'.
Add ylim for '*dsi' Tplot Variables of datatype = 'Espin'.
Add processing for maximum and minimum of ylim, which consider trange.
The purpose of these processing is adjust the output.
The cause of ylim processing is that pytplot.timespan() doesn not effect in ylim.
These processing for ylim may be no need in future.